requirements.txt
# Web scraping
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
scrapy==2.11.1

# HTML parsing and article extraction
newspaper3k==0.2.8
python-dateutil==2.8.2

# Database
# SQLite is built into Python, no package needed

# Scheduling
schedule==1.2.1
APScheduler==3.10.4

# Utilities
python-dotenv==1.0.1
tqdm==4.66.1

# Optional: for JavaScript-heavy sites
# selenium==4.16.0
# playwright==1.41.0

# Optional: for web UI monitoring
# flask==3.0.0
